+++ src/SettingsLayout.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link, useLocation } from "react-router-dom";
-import Nav from "./components/nav.js";
-import Head from "./components/head.js";
-import Obfuscate from "./components/obfuscate.js";
+import Nav from "./components/nav.jsx";
+import Head from "./components/head.jsx";
+import Obfuscate from "./components/obfuscate.jsx";
 import clsx from "clsx";
 import { useTranslation } from 'react-i18next';