+++ src/components/obfuscate.jsx
@@ -4,9 +4,9 @@
 import clsx from "clsx";
 import { create } from "random-seed";
 import { memo } from "react";
-import { obfuscation } from "../consts.js";
+import { obfuscation } from "../consts.jsx";
 
-const rand = create(navigator.userAgent + global.location.origin);
+const rand = create(navigator.userAgent + window.location.origin);
 
 const chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
@@ -58,7 +58,7 @@ export function ObfuscateLayout() {
 
 class ObfuscateContext {
   constructor(seed) {
-    this.rand = create(seed + navigator.userAgent + global.location.origin);
+    this.rand = create(seed + navigator.userAgent + window.location.origin);
   }
   ellipsisClass() {
     return ellipsisClasses[this.rand(ellipsisClasses.length)];