+++ src/consts.jsx
@@ -1,6 +1,6 @@
 const obfuscation = true;
 const github = "https://github.com/Metallic-Web/Metallic";
 const discord = "https://discord.com/invite/tS29NAyE3H";
-const bareServerURL = new URL("/bare/", global.location);
+const bareServerURL = new URL("/bare/", window.location);
 
 export { obfuscation, github, discord, bareServerURL };