+++ src/i18n.jsx
@@ -1,6 +1,6 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
-import languages from "./locales/index.js";
+import languages from "./locales/index.jsx";
 
 var language = localStorage.getItem("language") || "en"