+++ src/locales/index.js
@@ -1,9 +1,0 @@
-import en from "./en/index.js"
-import es from "./es/index.js"
-import fr from "./fr/index.js"
-import zh from "./zh/index.js"
-import de from "./de/index.js"
-
-var languages = {en, es, fr, zh, de}
-
-export default languages;