+++ src/pages/apps.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import Nav from "../components/nav.js";
-import Head from "../components/head.js";
-import Obfuscate from "../components/obfuscate.js";
-import Proxy from "../components/proxy.js";
+import Nav from "../components/nav.jsx";
+import Head from "../components/head.jsx";
+import Obfuscate from "../components/obfuscate.jsx";
+import Proxy from "../components/proxy.jsx";
 import globeSVG from "../assets/globe.svg";
-import gitems from "../apps.js";
-import { getLink } from "../util.js";
+import gitems from "../apps.jsx";
+import { getLink } from "../util.jsx";
 
 function Apps() {
   var proxy = React.useRef();